:truck: Rename(boardList): boardList 공통 컴포넌트로 변경
activity->공통
